Moves $validationCallback property from BBCode to GenericBBCode
